Use the `cargo-lock` crate
An extraction from this crate:

https://github.com/RustSec/cargo-lock
