+++ src/vulnerability.rs
@@ -44,7 +44,7 @@ impl Vulnerability {
                 for (path, versions) in &affected.functions {
                     if versions
                         .iter()
-                        .any(|req| req.matches(&self.package.version))
+                        .any(|req| req.matches(&self.package.version.clone().into()))
                     {
                         result.push(path.clone());
                     }
